+++ src/components/Popover/Popover.stories.tsx
@@ -1,5 +1,6 @@
 import React, { useState } from 'react';
 import { storiesOf } from '@storybook/react';
+import { PopoverAlign, PopoverPosition, PopoverTrigger } from '../../types';
 import { Icons } from '../Icons';
 import { Popover } from './Popover';
 import { Button } from '../Button';
@@ -19,10 +20,10 @@ storiesOf('Popover', module).add('Default', () => {
     { label: 'click', value: 'click' },
   ];
 
-  const selectedValuesExample = [];
-
-  const [newSelectedTrigger, setNewSelectedTrigger] = useState<'hover' | 'click'>('hover');
-  const [selectedValues, setSelectedValues] = useState(selectedValuesExample);
+  const [newSelectedTrigger, setNewSelectedTrigger] = useState<PopoverTrigger>(
+    PopoverTrigger.Hover
+  );
+  const [selectedValues, setSelectedValues] = useState([]);
   const [newColor, setNewColor] = useState('#000');
   const [newWidth, setNewWidth] = useState('400px');
 
@@ -43,10 +44,10 @@ storiesOf('Popover', module).add('Default', () => {
   };
 
   const customDivStyle: React.CSSProperties = {
-    backgroundColor: '#3e7bfa',
+    backgroundColor: '#3E7BFA',
     borderRadius: '4px',
     padding: '8px 4px',
-    color: '#fff',
+    color: '#FFF',
     height: '48px',
     width: '132px',
     textAlign: 'center',
@@ -79,7 +80,7 @@ storiesOf('Popover', module).add('Default', () => {
       <p>Ways to trigger with custom HTML</p>
       <div style={wrapperStyleLessBlocks}>
         <Popover
-          trigger='hover'
+          trigger={PopoverTrigger.Hover}
           content={
             <>
               <p>Arrow icons list</p>
@@ -107,7 +108,7 @@ storiesOf('Popover', module).add('Default', () => {
         </Popover>
 
         <Popover
-          trigger='click'
+          trigger={PopoverTrigger.Click}
           content={
             <>
               <p>Checkbox group</p>
@@ -136,7 +137,7 @@ storiesOf('Popover', module).add('Default', () => {
       <div style={wrapperStyle}>
         <Popover
           trigger={newSelectedTrigger}
-          position='right'
+          position={PopoverPosition.Right}
           content={<div>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</div>}
         >
           <div style={customDivStyle}>
@@ -146,7 +147,7 @@ storiesOf('Popover', module).add('Default', () => {
         </Popover>
         <Popover
           trigger={newSelectedTrigger}
-          position='top'
+          position={PopoverPosition.Top}
           content={<div>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</div>}
         >
           <div style={customDivStyle}>
@@ -156,7 +157,7 @@ storiesOf('Popover', module).add('Default', () => {
         </Popover>
         <Popover
           trigger={newSelectedTrigger}
-          position='bottom'
+          position={PopoverPosition.Bottom}
           content={<div>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</div>}
         >
           <div style={customDivStyle}>
@@ -167,7 +168,7 @@ storiesOf('Popover', module).add('Default', () => {
         </Popover>
         <Popover
           trigger={newSelectedTrigger}
-          position='left'
+          position={PopoverPosition.Left}
           content={<div>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</div>}
         >
           <div style={customDivStyle}>
@@ -181,8 +182,8 @@ storiesOf('Popover', module).add('Default', () => {
       <div style={wrapperStyle}>
         <Popover
           trigger={newSelectedTrigger}
-          position='right'
-          align='start'
+          position={PopoverPosition.Right}
+          align={PopoverAlign.Start}
           content={<div>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</div>}
         >
           <div style={customDivStyle}>
@@ -192,8 +193,8 @@ storiesOf('Popover', module).add('Default', () => {
         </Popover>
         <Popover
           trigger={newSelectedTrigger}
-          position='top'
-          align='start'
+          position={PopoverPosition.Top}
+          align={PopoverAlign.Start}
           content={<div>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</div>}
         >
           <div style={customDivStyle}>
@@ -202,8 +203,8 @@ storiesOf('Popover', module).add('Default', () => {
         </Popover>
         <Popover
           trigger={newSelectedTrigger}
-          position='bottom'
-          align='start'
+          position={PopoverPosition.Bottom}
+          align={PopoverAlign.Start}
           content={<div>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</div>}
         >
           <div style={customDivStyle}>
@@ -212,8 +213,8 @@ storiesOf('Popover', module).add('Default', () => {
         </Popover>
         <Popover
           trigger={newSelectedTrigger}
-          position='left'
-          align='start'
+          position={PopoverPosition.Left}
+          align={PopoverAlign.Start}
           content={<div>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</div>}
         >
           <div style={customDivStyle}>
@@ -226,8 +227,8 @@ storiesOf('Popover', module).add('Default', () => {
       <div style={wrapperStyle}>
         <Popover
           trigger={newSelectedTrigger}
-          position='right'
-          align='end'
+          position={PopoverPosition.Right}
+          align={PopoverAlign.End}
           content={<div>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</div>}
         >
           <div style={customDivStyle}>
@@ -236,8 +237,8 @@ storiesOf('Popover', module).add('Default', () => {
         </Popover>
         <Popover
           trigger={newSelectedTrigger}
-          position='top'
-          align='end'
+          position={PopoverPosition.Top}
+          align={PopoverAlign.End}
           content={<div>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</div>}
         >
           <div style={customDivStyle}>
@@ -247,8 +248,8 @@ storiesOf('Popover', module).add('Default', () => {
         </Popover>
         <Popover
           trigger={newSelectedTrigger}
-          position='bottom'
-          align='end'
+          position={PopoverPosition.Bottom}
+          align={PopoverAlign.End}
           content={<div>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</div>}
         >
           <div style={customDivStyle}>
@@ -258,8 +259,8 @@ storiesOf('Popover', module).add('Default', () => {
         </Popover>
         <Popover
           trigger={newSelectedTrigger}
-          position='left'
-          align='end'
+          position={PopoverPosition.Left}
+          align={PopoverAlign.End}
           content={<div>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</div>}
         >
           <div style={customDivStyle}>
@@ -273,7 +274,7 @@ storiesOf('Popover', module).add('Default', () => {
       <div style={wrapperStyle}>
         <Popover
           trigger={newSelectedTrigger}
-          position='bottom'
+          position={PopoverPosition.Bottom}
           content={<div>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</div>}
         >
           <Button>Offset=0</Button>
@@ -282,7 +283,7 @@ storiesOf('Popover', module).add('Default', () => {
         <Popover
           offset={8}
           trigger={newSelectedTrigger}
-          position='bottom'
+          position={PopoverPosition.Bottom}
           content={<div>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</div>}
         >
           <Button>Offset=8</Button>
@@ -291,7 +292,7 @@ storiesOf('Popover', module).add('Default', () => {
         <Popover
           offset={16}
           trigger={newSelectedTrigger}
-          position='bottom'
+          position={PopoverPosition.Bottom}
           content={<div>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</div>}
         >
           <Button>Offset=16</Button>
@@ -300,7 +301,7 @@ storiesOf('Popover', module).add('Default', () => {
         <Popover
           offset={24}
           trigger={newSelectedTrigger}
-          position='bottom'
+          position={PopoverPosition.Bottom}
           content={<div>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</div>}
         >
           <Button>Offset=24</Button>
@@ -312,7 +313,7 @@ storiesOf('Popover', module).add('Default', () => {
         <Popover
           hasArrow={false}
           trigger={newSelectedTrigger}
-          position='bottom'
+          position={PopoverPosition.Bottom}
           content={<div>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</div>}
         >
           <Button>Offset=0</Button>
@@ -322,7 +323,7 @@ storiesOf('Popover', module).add('Default', () => {
           offset={8}
           hasArrow={false}
           trigger={newSelectedTrigger}
-          position='bottom'
+          position={PopoverPosition.Bottom}
           content={<div>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</div>}
         >
           <Button>Offset=8</Button>
@@ -332,7 +333,7 @@ storiesOf('Popover', module).add('Default', () => {
           offset={16}
           hasArrow={false}
           trigger={newSelectedTrigger}
-          position='bottom'
+          position={PopoverPosition.Bottom}
           content={<div>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</div>}
         >
           <Button>Offset=16</Button>
@@ -342,7 +343,7 @@ storiesOf('Popover', module).add('Default', () => {
           offset={24}
           hasArrow={false}
           trigger={newSelectedTrigger}
-          position='bottom'
+          position={PopoverPosition.Bottom}
           content={<div>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</div>}
         >
           <Button>Offset=24</Button>